REF: move PaybackGuaranteeType to credit module
